tests/wpt/tests/html/webappapis/dynamic-markup-insertion/document-write/041.html

<!doctype html>
<title>document.write entity</title>
<script src="/resources/testharness.js"></script><script src="/resources/testharnessreport.js"></script>
<script>
var t = test(function() {
  var s = "<body><span>&notin;abc";
  for (var i=0; i<s.length; i++) {
    document.write(s[i]);
  }
  assert_equals(document.body.childNodes[0].textContent, "\u2209abc");
});
</script>
<div id="log"></div>
